+++ trunk/src/psxyz.c
@@ -1037,7 +1037,7 @@ int GMT_psxyz (void *V_API, int mode, void *args)
 					GMT_plane_perspective (GMT, GMT_Z + GMT_ZW, GMT->current.proj.z_level);
 					if ((GMT->current.plot.n = GMT_geo_to_xy_line (GMT, L->coord[GMT_X], L->coord[GMT_Y], L->n_rows)) == 0) continue;
 					GMT_setfill (GMT, &current_fill, outline_active);
-					GMT_draw_front (GMT, GMT->current.plot.x, GMT->current.plot.y, (unsigned int)GMT->current.plot.n, &S.f);
+					GMT_draw_front (GMT, GMT->current.plot.x, GMT->current.plot.y, GMT->current.plot.n, &S.f);
 				}
 
 				GMT_free (GMT, xp);
